woocommerce update

+++ wp-content/plugins/woocommerce/includes/data-stores/class-wc-shipping-zone-data-store.php
@@ -14,7 +14,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  *
  * @version  3.0.0
  */
-class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shipping_Zone_Data_Store_Interface, WC_Object_Data_Store_Interface {
+class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Object_Data_Store_Interface, WC_Shipping_Zone_Data_Store_Interface {
 
 	/**
 	 * Method to create a new shipping zone.
@@ -25,7 +25,8 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 	public function create( &$zone ) {
 		global $wpdb;
 		$wpdb->insert(
-			$wpdb->prefix . 'woocommerce_shipping_zones', array(
+			$wpdb->prefix . 'woocommerce_shipping_zones',
+			array(
 				'zone_name'  => $zone->get_zone_name(),
 				'zone_order' => $zone->get_zone_order(),
 			)
@@ -34,7 +35,7 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 		$zone->save_meta_data();
 		$this->save_locations( $zone );
 		$zone->apply_changes();
-		WC_Cache_Helper::incr_cache_prefix( 'shipping_zones' );
+		WC_Cache_Helper::invalidate_cache_group( 'shipping_zones' );
 		WC_Cache_Helper::get_transient_version( 'shipping', true );
 	}
 
@@ -48,16 +49,18 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 		global $wpdb;
 		if ( $zone->get_id() ) {
 			$wpdb->update(
-				$wpdb->prefix . 'woocommerce_shipping_zones', array(
+				$wpdb->prefix . 'woocommerce_shipping_zones',
+				array(
 					'zone_name'  => $zone->get_zone_name(),
 					'zone_order' => $zone->get_zone_order(),
-				), array( 'zone_id' => $zone->get_id() )
+				),
+				array( 'zone_id' => $zone->get_id() )
 			);
 		}
 		$zone->save_meta_data();
 		$this->save_locations( $zone );
 		$zone->apply_changes();
-		WC_Cache_Helper::incr_cache_prefix( 'shipping_zones' );
+		WC_Cache_Helper::invalidate_cache_group( 'shipping_zones' );
 		WC_Cache_Helper::get_transient_version( 'shipping', true );
 	}
 
@@ -71,33 +74,41 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 	public function read( &$zone ) {
 		global $wpdb;
 
-		$zone_data = false;
-
-		if ( 0 !== $zone->get_id() || '0' !== $zone->get_id() ) {
-			$zone_data = $wpdb->get_row(
-				$wpdb->prepare(
-					"SELECT zone_name, zone_order FROM {$wpdb->prefix}woocommerce_shipping_zones WHERE zone_id = %d LIMIT 1",
-					$zone->get_id()
-				)
-			);
-		}
-
+		// Zone 0 is used as a default if no other zones fit.
 		if ( 0 === $zone->get_id() || '0' === $zone->get_id() ) {
 			$this->read_zone_locations( $zone );
 			$zone->set_zone_name( __( 'Locations not covered by your other zones', 'woocommerce' ) );
 			$zone->read_meta_data();
 			$zone->set_object_read( true );
+
+			/**
+			 * Indicate that the WooCommerce shipping zone has been loaded.
+			 *
+			 * @param WC_Shipping_Zone $zone The shipping zone that has been loaded.
+			 */
 			do_action( 'woocommerce_shipping_zone_loaded', $zone );
-		} elseif ( $zone_data ) {
-			$zone->set_zone_name( $zone_data->zone_name );
-			$zone->set_zone_order( $zone_data->zone_order );
-			$this->read_zone_locations( $zone );
-			$zone->read_meta_data();
-			$zone->set_object_read( true );
-			do_action( 'woocommerce_shipping_zone_loaded', $zone );
-		} else {
+			return;
+		}
+
+		$zone_data = $wpdb->get_row(
+			$wpdb->prepare(
+				"SELECT zone_name, zone_order FROM {$wpdb->prefix}woocommerce_shipping_zones WHERE zone_id = %d LIMIT 1",
+				$zone->get_id()
+			)
+		);
+
+		if ( ! $zone_data ) {
 			throw new Exception( __( 'Invalid data store.', 'woocommerce' ) );
 		}
+
+		$zone->set_zone_name( $zone_data->zone_name );
+		$zone->set_zone_order( $zone_data->zone_order );
+		$this->read_zone_locations( $zone );
+		$zone->read_meta_data();
+		$zone->set_object_read( true );
+
+		/** This action is documented in includes/datastores/class-wc-shipping-zone-data-store.php. */
+		do_action( 'woocommerce_shipping_zone_loaded', $zone );
 	}
 
 	/**
@@ -109,17 +120,30 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 	 * @return void
 	 */
 	public function delete( &$zone, $args = array() ) {
-		if ( $zone->get_id() ) {
+		$zone_id = $zone->get_id();
+
+		if ( $zone_id ) {
 			global $wpdb;
-			$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zone_methods', array( 'zone_id' => $zone->get_id() ) );
-			$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zone_locations', array( 'zone_id' => $zone->get_id() ) );
-			$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zones', array( 'zone_id' => $zone->get_id() ) );
-			WC_Cache_Helper::incr_cache_prefix( 'shipping_zones' );
-			$id = $zone->get_id();
+
+			// Delete methods and their settings.
+			$methods = $this->get_methods( $zone_id, false );
+
+			if ( $methods ) {
+				foreach ( $methods as $method ) {
+					$this->delete_method( $method->instance_id );
+				}
+			}
+
+			// Delete zone.
+			$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zone_locations', array( 'zone_id' => $zone_id ) );
+			$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zones', array( 'zone_id' => $zone_id ) );
+
 			$zone->set_id( null );
-			WC_Cache_Helper::incr_cache_prefix( 'shipping_zones' );
+
+			WC_Cache_Helper::invalidate_cache_group( 'shipping_zones' );
 			WC_Cache_Helper::get_transient_version( 'shipping', true );
-			do_action( 'woocommerce_delete_shipping_zone', $id );
+
+			do_action( 'woocommerce_delete_shipping_zone', $zone_id );
 		}
 	}
 
@@ -140,7 +164,7 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 			$raw_methods_sql = "SELECT method_id, method_order, instance_id, is_enabled FROM {$wpdb->prefix}woocommerce_shipping_zone_methods WHERE zone_id = %d";
 		}
 
-		return $wpdb->get_results( $wpdb->prepare( $raw_methods_sql, $zone_id ) ); // phpcs:ignore WordPress.WP.PreparedSQL.NotPrepared
+		return $wpdb->get_results( $wpdb->prepare( $raw_methods_sql, $zone_id ) ); // phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared
 	}
 
 	/**
@@ -190,7 +214,17 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 	 */
 	public function delete_method( $instance_id ) {
 		global $wpdb;
+
+		$method = $this->get_method( $instance_id );
+
+		if ( ! $method ) {
+			return;
+		}
+
+		delete_option( 'woocommerce_' . $method->method_id . '_' . $instance_id . '_settings' );
+
 		$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zone_methods', array( 'instance_id' => $instance_id ) );
+
 		do_action( 'woocommerce_delete_shipping_zone_method', $instance_id );
 	}
 
@@ -241,12 +275,22 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 			}
 		}
 
+		/**
+		 * Get shipping zone criteria
+		 *
+		 * @since 3.6.6
+		 * @param array $criteria Get zone criteria.
+		 * @param array $package Package information.
+		 * @param array $postcode_locations Postcode range and wildcard matching.
+		 */
+		$criteria = apply_filters( 'woocommerce_get_zone_criteria', $criteria, $package, $postcode_locations );
+
 		// Get matching zones.
 		return $wpdb->get_var(
 			"SELECT zones.zone_id FROM {$wpdb->prefix}woocommerce_shipping_zones as zones
 			LEFT OUTER JOIN {$wpdb->prefix}woocommerce_shipping_zone_locations as locations ON zones.zone_id = locations.zone_id AND location_type != 'postcode'
-			WHERE " . implode( ' ', $criteria ) // phpcs:ignore WordPress.WP.PreparedSQL.NotPrepared
-			. ' ORDER BY zone_order ASC, zone_id ASC LIMIT 1'
+			WHERE " . implode( ' ', $criteria ) // phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared
+			. ' ORDER BY zone_order ASC, zones.zone_id ASC LIMIT 1'
 		);
 	}
 
@@ -266,7 +310,7 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 	 * Return a zone ID from an instance ID.
 	 *
 	 * @since  3.0.0
-	 * @param  int $id Instnace ID.
+	 * @param  int $id Instance ID.
 	 * @return int
 	 */
 	public function get_zone_id_by_instance_id( $id ) {
@@ -317,7 +361,8 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 
 		foreach ( $zone->get_zone_locations( 'edit' ) as $location ) {
 			$wpdb->insert(
-				$wpdb->prefix . 'woocommerce_shipping_zone_locations', array(
+				$wpdb->prefix . 'woocommerce_shipping_zone_locations',
+				array(
 					'zone_id'       => $zone->get_id(),
 					'location_code' => $location->code,
 					'location_type' => $location->type,
